EditExpenseActivity - added delete button on action bar

+++ app/src/main/java/com/example/fico/presentation/helper/SwipeToDeleteCallback.kt
@@ -1,4 +1,3 @@
-
 import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.lifecycle.LifecycleOwner
@@ -13,7 +12,12 @@ import com.google.android.material.snackbar.Snackbar
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
-class SwipeToDeleteCallback(private val recyclerView: RecyclerView, private val viewModel: ExpenseListViewModel, private val adapter: ExpenseListAdapter, private val lifecycleOwner : LifecycleOwner) :
+class SwipeToDeleteCallback(
+    private val recyclerView: RecyclerView,
+    private val viewModel: ExpenseListViewModel,
+    private val adapter: ExpenseListAdapter,
+    private val lifecycleOwner: LifecycleOwner
+) :
     ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT) {
 
     override fun onMove(
@@ -27,12 +31,17 @@ class SwipeToDeleteCallback(private val recyclerView: RecyclerView, private val 
 
     @RequiresApi(Build.VERSION_CODES.O)
     override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-            //Format data to delete item
-            val position = viewHolder.adapterPosition
-            val deleteItem = adapter.getDataAtPosition(position)
+        //Format data to delete item
+        val position = viewHolder.adapterPosition
+        val deleteItem = adapter.getDataAtPosition(position)
+
+        //Check if expense in not installment
+        if (deleteItem.id.length < 41) {
             val formattedPaymentDate = FormatValuesToDatabase().expenseDate(deleteItem.paymentDate)
-            val formattedPurchaseDate = FormatValuesToDatabase().expenseDate(deleteItem.purchaseDate)
-            val formattedInputDateTime = FormatValuesToDatabase().expenseDate(deleteItem.inputDateTime)
+            val formattedPurchaseDate =
+                FormatValuesToDatabase().expenseDate(deleteItem.purchaseDate)
+            val formattedInputDateTime =
+                FormatValuesToDatabase().expenseDate(deleteItem.inputDateTime)
             val expencePrice = "-${deleteItem.price.replace("R$ ", "").replace(",", ".")}"
             val deleteItemFormatted = Expense(
                 deleteItem.id,
@@ -43,9 +52,12 @@ class SwipeToDeleteCallback(private val recyclerView: RecyclerView, private val 
                 formattedPurchaseDate,
                 formattedInputDateTime
             )
-
             //Delete Item and update expense list
             viewModel.deleteExpense(deleteItemFormatted)
+        }else{
+            viewModel.onInstallmentExpenseSwiped()
+        }
+
 
     }
 }
